+++ buession-redis/src/main/java/com/buession/redis/client/lettuce/operations/AbstractStreamOperations.java
@@ -35,9 +35,6 @@ import com.buession.redis.core.StreamFull;
 import com.buession.redis.core.StreamGroup;
 import com.buession.redis.core.StreamPending;
 import com.buession.redis.core.StreamPendingSummary;
-import com.buession.redis.core.command.args.XAddArgument;
-import com.buession.redis.core.command.args.XClaimArgument;
-import com.buession.redis.core.command.args.XTrimArgument;
 import com.buession.redis.core.internal.convert.Converters;
 import com.buession.redis.utils.SafeEncoder;
 
@@ -155,7 +152,7 @@ public abstract class AbstractStreamOperations<C extends LettuceRedisClient> ext
 	}
 
 	@Override
-	public StreamFull xInfoStream(final String key, final boolean full, final int count) {
+	public StreamFull xInfoStream(final String key, final boolean full, final long count) {
 		return xInfoStream(SafeEncoder.encode(key), full, count);
 	}
 
@@ -176,7 +173,7 @@ public abstract class AbstractStreamOperations<C extends LettuceRedisClient> ext
 
 	@Override
 	public List<StreamPending> xPending(final String key, final String groupName, final StreamEntryId start,
-										final StreamEntryId end, final int count) {
+										final StreamEntryId end, final long count) {
 		return xPending(SafeEncoder.encode(key), SafeEncoder.encode(groupName), start, end, count);
 	}
 
@@ -187,7 +184,7 @@ public abstract class AbstractStreamOperations<C extends LettuceRedisClient> ext
 
 	@Override
 	public List<StreamPending> xPending(final String key, final String groupName, final long minIdleTime,
-										final StreamEntryId start, final StreamEntryId end, final int count) {
+										final StreamEntryId start, final StreamEntryId end, final long count) {
 		return xPending(SafeEncoder.encode(key), SafeEncoder.encode(groupName), minIdleTime, start, end, count);
 	}
 
@@ -200,14 +197,14 @@ public abstract class AbstractStreamOperations<C extends LettuceRedisClient> ext
 
 	@Override
 	public List<StreamPending> xPending(final String key, final String groupName, final StreamEntryId start,
-										final StreamEntryId end, final int count, final String consumerName) {
+										final StreamEntryId end, final long count, final String consumerName) {
 		return xPending(SafeEncoder.encode(key), SafeEncoder.encode(groupName), start, end, count,
 				SafeEncoder.encode(consumerName));
 	}
 
 	@Override
 	public List<StreamPending> xPending(final String key, final String groupName, final long minIdleTime,
-										final StreamEntryId start, final StreamEntryId end, final int count,
+										final StreamEntryId start, final StreamEntryId end, final long count,
 										final String consumerName) {
 		return xPending(SafeEncoder.encode(key), SafeEncoder.encode(groupName), minIdleTime, start, end, count,
 				SafeEncoder.encode(consumerName));
@@ -220,7 +217,7 @@ public abstract class AbstractStreamOperations<C extends LettuceRedisClient> ext
 
 	@Override
 	public List<StreamEntry> xRange(final String key, final StreamEntryId start, final StreamEntryId end,
-									final int count) {
+									final long count) {
 		return xRange(SafeEncoder.encode(key), start, end, count);
 	}
 
@@ -231,7 +228,7 @@ public abstract class AbstractStreamOperations<C extends LettuceRedisClient> ext
 
 	@Override
 	public List<StreamEntry> xRevRange(final String key, final StreamEntryId end, final StreamEntryId start,
-									   final int count) {
+									   final long count) {
 		return xRevRange(SafeEncoder.encode(key), end, start, count);
 	}
